+++ apps/scoutgame/components/claim/components/PointsTable/PointsReceiptRewardRow.tsx
@@ -22,7 +22,12 @@ function BuilderRewardRow({ builderReward }: { builderReward: BuilderPointsRecei
     <>
       <DividerRow />
       <TableRow key={`${builderReward.weekNumber}builder-rewards`}>
-        <TableCell align='left'>
+        <TableCell
+          align='left'
+          sx={{
+            minWidth: 150
+          }}
+        >
           <Typography>Builder rewards</Typography>
         </TableCell>
         <TableCell align='center'>
@@ -45,7 +50,12 @@ function LeaderboardRankRewardRow({
     <>
       <DividerRow />
       <TableRow key={`${leaderboardRankReward.weekNumber}rank`}>
-        <TableCell align='left'>
+        <TableCell
+          align='left'
+          sx={{
+            minWidth: 150
+          }}
+        >
           <Typography>Finished {getOrdinal(leaderboardRankReward.rank)}</Typography>
         </TableCell>
         <TableCell align='center'>
@@ -64,7 +74,12 @@ function SoldNftsRewardRow({ soldNftsReward }: { soldNftsReward: SoldNftsPointsR
     <>
       <DividerRow />
       <TableRow key={`${soldNftsReward.weekNumber}sold-nft`}>
-        <TableCell align='left'>
+        <TableCell
+          align='left'
+          sx={{
+            minWidth: 150
+          }}
+        >
           <Stack direction='row' alignItems='center' justifyContent='flex-start' gap={0.5}>
             <Typography>Sold {soldNftsReward.quantity}</Typography>
             <Image alt='card' src='/images/profile/icons/card.svg' width={18} height={18} />
