+++ packages/scoutgame/src/points/getClaimablePoints.ts
@@ -2,27 +2,43 @@ import { prisma } from '@charmverse/core/prisma-client';
 
 import { currentSeason, getPreviousSeason } from '../dates';
 
-export async function getClaimablePoints({ userId, week }: { week?: string; userId: string }): Promise<number> {
+export async function getClaimablePoints({ userId, week }: { week?: string; userId: string }): Promise<{
+  points: number;
+  bonusPartners: string[];
+}> {
   const previousSeason = getPreviousSeason(currentSeason);
-  const claimableSeasons = [previousSeason, currentSeason].filter(Boolean);
-  if (claimableSeasons.length === 0) {
+  const seasons = [previousSeason, currentSeason].filter(Boolean);
+  if (seasons.length === 0) {
     throw new Error(`No seasons found to claim points: ${currentSeason}`);
   }
   const pointsReceipts = await prisma.pointsReceipt.findMany({
     where: {
       recipientId: userId,
-      claimedAt: null,
+      claimedAt: { equals: null },
       event: {
         week,
         season: {
-          in: claimableSeasons
+          in: seasons
         }
       }
     },
     select: {
-      value: true
+      value: true,
+      event: {
+        select: {
+          bonusPartner: true
+        }
+      }
     }
   });
 
-  return pointsReceipts.reduce((acc, receipt) => acc + receipt.value, 0);
+  const totalUnclaimedPoints = pointsReceipts.reduce((acc, receipt) => acc + receipt.value, 0);
+  const bonusPartners = Array.from(new Set(pointsReceipts.map((receipt) => receipt.event.bonusPartner))).filter(
+    (bp) => bp !== null
+  );
+
+  return {
+    points: totalUnclaimedPoints,
+    bonusPartners
+  };
 }
